// src/app/location/models/province.model.ts
export interface ProvinceModel {
    id: number;
    label: string;
    regionId: number;
    cornAvg: number;
    irrigated1stq: number;
    irrigated2ndq: number;
    irrigated3rdq: number;
    irrigated4thq: number;
    irrigatedDry: number;
    irrigatedWet: number;
    rainfed1stq: number;
    rainfed2ndq: number;
    rainfed3rdq: number;
    rainfed4thq: number;
    rainfedDry: number;
    rainfedWet: number;
}